requirements.txt
requests==2.28.1
line-bot-sdk==2.3.0
Flask==2.2.2
google-cloud-vision==3.3.0
